+++ lib/screens/reminders.dart
@@ -524,7 +524,7 @@ class _RemindersState extends State<Reminders> {
             SingleChildScrollView(
               child: Padding(
                 padding: EdgeInsets.symmetric(
-                  horizontal: Responsive.width(context, 50),
+                  horizontal: Responsive.centeredHorizontalPadding(context, 50),
                   vertical: Responsive.height(context, 24),
                 ),
                 child: Column(