+++ lib/screens/settings/personal_preferences.dart
@@ -576,7 +576,7 @@ class _PersonalPreferencesState extends State<PersonalPreferences> {
             SingleChildScrollView(
               child: Padding(
                 padding: EdgeInsets.symmetric(
-                  horizontal: Responsive.width(context, 50),
+                  horizontal: Responsive.centeredHorizontalPadding(context, 50),
                   vertical: Responsive.height(context, 24),
                 ),
                 child: Column(